removed formula functionality
there was a formula button that I intended to allow the addition of functions on the values of each cells. the drop down would have things like "celsius" as an option which would then convert the value in that cell to celsius. I was informed that the correct values should be coming in anyways so I abandoned this idea for now.