added a single cell edit function
there is now a button in the lower right hand corner of the screen that when clicked enables single cells to be clicked. once clicked, the single cells toggle a window that allows for the title and label of that cell to be editted. Currently there is no way to save the changes, however. I also added another image/license to the sources.md file.