fixed bug with modifying cells
when modifying cells with the master edit button it would make cells unmodifiable with the modify table button afterwards. I looked into it and found that the master edit button method was editing the text of the div instead of the paragraph child. This resulted in teh paragraph child being deleted, thus making it untargettable by modify table.