Added camera timer
I added a bit of logic that checks to see if the url for the camera image has changed. if it has the camera timer resets to zero. if not, it keeps counting upwards every second and displays in the status bar at the bottom. Next change is to create a status bar that spans the entire width of the page.