slowly getting rid of control panel
I am in the process of phasing out the control panel which operates on a "modify-all" functionality. I am moving towards creating a "modify-per-element" control scheme. Currently the user can click on the edit icon in the lower right-hand corner, click on a cell (which pops open an edit-pane, and then the title, label and font size can all be editted live. The next improvement will be the ability to hide cells. Eventually I would like this functionality to work for all elements on the page.