added the ability to hide cells
I added a hide button to the edit window div that will change the opacity of the cell to .2 until the edit button in the lower right is clicked, at which point the hidden element will actually be hidden. If the edit button is clicked once more, the element will appear with reduced opacity until the unhide button is clicked.