cam images are now being loaded in to divs
I got it so the function will load images into background images on divs. These divs have a display none property, so the image will not load until the user specifies. This is because the image will be loaded with css. The next step is to create a selection list that the user can choose images from.