user can now dynamically add cam feeds
I added a selection list that has options featurning a single cam url and a single cam name. when a user makes a selection, they can hit the create cam button and that image should be made visible. The user can then drag and resize to their desire. next step is to allow these images to preload so that we do not get scan lines on them.