added tree view
there is now a file tree that represents the data coming in from the cameras. currenlty the user can drag leaf-nodes from the file tree to the main canvas and it will drop the title. my goal is to change this so that it will create a cell that updates with data.