iterateStations now checks for title,units,value
the function currently checks to see if there is a 'title' property. if there is, it adds a 'title' and 'units' properties to the tree object. It also changes the path property so that it includes the value property in the path. Currently it is set up so that both the new and old formats work. the old format being an object w/o title, value, and unit properties.