fixed camera images creation
when the structure of the object changed, it broke camera image creation. the function now looks at [station][cameras] and loops through children of [cameras] finding properties labeled [image_url]. it then adds the values of these children to the array. It should be able to find multiple cameras under the same station now. My goal is to change it so that it can load them from anywhere in the nested object.